Move start of transaction for the home to the app.tsx
